Refactor service creation
